requirements.txt
earthengine_api==0.1.342
ffmpeg_python==0.2.0
jsonargparse==4.20.0
matplotlib==3.5.3
natsort==8.2.0
numpy==1.21.6
pandas==1.3.5
pydantic==1.10.5
PyYAML==6.0
rasterio==1.2.10
requests_futures==1.0.0
rioxarray==0.9.1
scikit_learn==1.0.2
seaborn==0.12.2
segmentation_models_pytorch==0.3.2
setuptools==65.6.3
torch==1.13.1
torchgeo==0.4.0
torchmetrics==0.11.3
torchvision==0.14.1
tqdm==4.64.1
wandb==0.13.10
xarray==0.20.2
xgboost==1.6.2